worker: move gouging to internal package
